Update y_AlphaSim.m
After observing a problem with AlphaSim for small masks in combination with high smoothness, I looked into this code and edited it. The most important point is the relocation of the application of the mask. Initially this was done before the the gauss filter but the gauss filter will blur the bounderies of the masked region which will cause problems later in the code.
A more detailed description of the edits can be found here:
http://rfmri.org/content/alphasim-problem-critical-bug